[Refactor]클래스 재구조화, 가독성 높이기
