--- src/GeneratNumber.java
@@ -1,3 +1,0 @@
-public interface GeneratNumber {
-    public int generateNumber();
-}
