+++ src/Main.java
@@ -1,12 +1,6 @@
-import java.util.ArrayList;
-import java.util.List;
-
 public class Main {
     public static void main(String[] args) {
-        PlayServiceImpl playService = new PlayServiceImpl();
-
-        playService.start();
-
-
+        PlayService playService = new PlayService();
+        playService.play();
     }
 }
